Security ready

+++ src/main/java/com/example/Todolist/controller/TaskController.java
@@ -1,19 +1,28 @@
 package com.example.Todolist.controller;
 
 import com.example.Todolist.model.Task;
+import com.example.Todolist.model.User;
 import com.example.Todolist.repository.TaskRepository;
+import com.example.Todolist.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import java.util.List;
+
 @Controller
 public class TaskController {
 
     @Autowired
     private TaskRepository taskRepository;
+
+    @Autowired
+    private UserRepository userRepository;
 
     /**
      * Returns every task to the user
@@ -22,8 +31,12 @@ public class TaskController {
      */
     @RequestMapping("/tasklist")
     public String taskList(Model model){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        org.springframework.security.core.userdetails.User userDetails = (org.springframework.security.core.userdetails.User) auth.getPrincipal();
+        User user = userRepository.findByUsername( userDetails.getUsername() );
+        List<Task> tasks = taskRepository.findByUser(user);
         model.addAttribute("task1", new Task());
-        model.addAttribute("tasks", taskRepository.findAll());
+        model.addAttribute("tasks", tasks);
         return "tasklist";
     }
 
@@ -46,6 +59,11 @@ public class TaskController {
      */
     @RequestMapping(value = "/save", method = RequestMethod.POST)
     public String save(Task task){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        org.springframework.security.core.userdetails.User userDetails = (org.springframework.security.core.userdetails.User) auth.getPrincipal();
+        User user = userRepository.findByUsername(userDetails.getUsername());
+
+        task.setUser(user);
         taskRepository.save(task);
 
         return "redirect:/tasklist";
@@ -58,7 +76,14 @@ public class TaskController {
      */
     @RequestMapping(value = "/delete/{id}", method = RequestMethod.GET)
     public  String deleteTask(@PathVariable("id") Long id, Model model){
-        taskRepository.delete(id);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        org.springframework.security.core.userdetails.User userDetails = (org.springframework.security.core.userdetails.User) auth.getPrincipal();
+        User user = userRepository.findByUsername( userDetails.getUsername() );
+        Task task = taskRepository.findOne(id);
+
+        if(task.getUser().equals(user)) taskRepository.delete(id);
+
+
         return "redirect:/tasklist";
     }
